+++ src/reflect/scala/reflect/internal/tpe/FindMembers.scala
@@ -55,7 +55,7 @@ trait FindMembers {
     private def searchConcreteThenDeferred: T = {
       val deferredSeen = walkBaseClasses(requiredFlags, excludedFlags | DEFERRED)
       if (deferredSeen) // OPT: the `if` avoids a second pass if the first pass didn't spot any candidates.
-        walkBaseClasses(requiredFlags | DEFERRED, excludedFlags & ~(DEFERRED.toLong))
+        walkBaseClasses(requiredFlags | DEFERRED, excludedFlags & ~DEFERRED)
       result
     }
